style: rename loader -> mqtt
