+++ agent.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const loader = require('./lib/loader');
+const mqtt = require('./lib/mqtt');
 module.exports = app => {
   if (app.config.mqtt.agent) {
-    loader(app);
+    mqtt(app);
   }
 };
